requirements.txt
langchain>=0.1.13
chromadb>=0.4.22

sentence-transformers>=2.2.2
PyMuPDF==1.22.5
streamlit>=1.32.0
python-dotenv>=1.0.1
pinecone-client 
langchain-pinecone
langchain-groq
